requirements.txt
textual>=0.40.0
python-dotenv>=1.0.0
gitpython>=3.1.0